Move ThreadLocalActiveSpan to .util
