+++ opentracing-util/src/main/java/io/opentracing/util/ThreadLocalActiveSpanSource.java
@@ -11,7 +11,12 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing;
+package io.opentracing.util;
+
+import io.opentracing.ActiveSpan;
+import io.opentracing.ActiveSpanSource;
+import io.opentracing.Span;
+import io.opentracing.Tracer;
 
 import java.util.concurrent.atomic.AtomicInteger;
